NavBar working dummy pages implemented
