+++ src/account.js
@@ -1,0 +1,13 @@
+import Container from 'react-bootstrap/Container';
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+import Button from 'react-bootstrap/Button';
+import React, { useState, Component } from 'react';
+import Form from 'react-bootstrap/Form';
+import Modal from 'react-bootstrap/Modal';
+
+export default class AccountHome extends Component {
+    render() {
+        return <Button>Account</Button>;
+    }
+}
